requirements.txt
streamlit==1.31.0
sentence-transformers
faiss-cpu==1.7.4
rank-bm25==0.2.2
PyPDF2==3.0.1
numpy==1.24.3
pandas==2.0.3
torch==2.1.2
transformers==4.36.2
scikit-learn==1.3.0
huggingface-hub
pytesseract
nltk
ollama
pymupdf